+++ src/test/java/ZUV/MiscValidatorTest.java
@@ -13,12 +13,25 @@ public class MiscValidatorTest extends ResourceCase  {
 		ZUGFeRDValidator zfv=new ZUGFeRDValidator();
 
 		String res=zfv.validate(null);
-		assertTrue(res.matches("<validation filename='' datetime='.*?'><messages><error type=\"10\">Filename not specified</error>\n" + 
-				"</messages><summary status='invalid'/></validation>"));
+		assertTrue(res.matches("<\\?xml version=\"1.0\" encoding=\"UTF-8\"\\?>\n" + 
+				"\n" + 
+				"<validation filename=\"\" datetime=\".*?\">\n" + 
+				"  <messages>\n" + 
+				"    <error type=\"10\">Filename not specified</error> \n" + 
+				"  </messages>\n" + 
+				"  <summary status=\"invalid\"/>\n" + 
+				"</validation>\n" + 
+				""));
 
 		res=zfv.validate("/dhfkbv/sfjkh");
-		assertTrue(res.matches("<validation filename='sfjkh' datetime='.*?'><messages><error type=\"1\">File not found</error>\n" + 
-								"</messages><summary status='invalid'/></validation>"));
+		assertTrue(res.matches("<\\?xml version=\"1.0\" encoding=\"UTF-8\"\\?>\n" + 
+				"\n" + 
+				"<validation filename=\"sfjkh\" datetime=\".*?\">\n" + 
+				"  <messages>\n" + 
+				"    <error type=\"1\">File not found</error> \n" + 
+				"  </messages>\n" + 
+				"  <summary status=\"invalid\"/>\n" + 
+				"</validation>\n"));
 
 		File tempFile=null;
 		try {
@@ -29,9 +42,15 @@ public class MiscValidatorTest extends ResourceCase  {
 		}
 
 		res=zfv.validate(tempFile.getAbsolutePath());
-		assertTrue(res.matches("<validation.*?><messages>" + 
-				"<error type=\"5\">File too small</error>\n" + 
-				"</messages><summary status='invalid'/></validation>"));
+		assertTrue(res.matches("<\\?xml version=\"1.0\" encoding=\"UTF-8\"\\?>\n" + 
+				"\n" + 
+				"<validation filename=\".*\" datetime=\".*\">\n" + 
+				"  <messages>\n" + 
+				"    <error type=\"5\">File too small</error> \n" + 
+				"  </messages>\n" + 
+				"  <summary status=\"invalid\"/>\n" + 
+				"</validation>\n" + 
+				""));
 		
 		
 		String fileContent = "ladhvkdbfk  wkhfbkhdhkb svbkfsvbksfbvk sdvsdvbksjdvbkfdsv sdvbskdvbsjhkvbfskh dvbskfvbkfsbvke"
@@ -49,8 +68,15 @@ public class MiscValidatorTest extends ResourceCase  {
 		 
 
 		res=zfv.validate(tempFile.getAbsolutePath());
-		assertTrue(res.matches("<validation filename='"+tempFile.getName()+"' .*?><messages><exception type=\"8\">File does not look like PDF nor XML .*</exception>\n" + 
-				"</messages><summary status='invalid'/></validation>"));
+		assertTrue(res.matches("<\\?xml version=\"1.0\" encoding=\"UTF-8\"\\?>\n" + 
+				"\n" + 
+				"<validation filename=\".*\" datetime=\".*\">\n" + 
+				"  <messages>\n" + 
+				"    <exception type=\"8\">File does not look like PDF nor XML \\(contains neither %PDF nor &lt;\\?xml\\)</exception> \n" + 
+				"  </messages>\n" + 
+				"  <summary status=\"invalid\"/>\n" + 
+				"</validation>\n" + 
+				""));
 
 		
 		// clean up